requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.13.1
idna==3.4
psycopg2-binary==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
